upgrade discordgo
