+++ pkg/discordutil/isession.go
@@ -1,0 +1,199 @@
+package discordutil
+
+import (
+	"image"
+	"io"
+	"time"
+
+	. "github.com/bwmarrin/discordgo"
+)
+
+type ISession interface {
+	AddHandler(handler interface{}) func()
+	AddHandlerOnce(handler interface{}) func()
+	Application(appID string) (st *Application, err error)
+	ApplicationAssets(appID string) (ass []*Asset, err error)
+	ApplicationBotCreate(appID string) (st *User, err error)
+	ApplicationCommand(appID, guildID, cmdID string) (cmd *ApplicationCommand, err error)
+	ApplicationCommandBulkOverwrite(appID string, guildID string, commands []*ApplicationCommand) (createdCommands []*ApplicationCommand, err error)
+	ApplicationCommandCreate(appID string, guildID string, cmd *ApplicationCommand) (ccmd *ApplicationCommand, err error)
+	ApplicationCommandDelete(appID, guildID, cmdID string) error
+	ApplicationCommandEdit(appID, guildID, cmdID string, cmd *ApplicationCommand) (updated *ApplicationCommand, err error)
+	ApplicationCommandPermissions(appID, guildID, cmdID string) (permissions *GuildApplicationCommandPermissions, err error)
+	ApplicationCommandPermissionsBatchEdit(appID, guildID string, permissions []*GuildApplicationCommandPermissions) (err error)
+	ApplicationCommandPermissionsEdit(appID, guildID, cmdID string, permissions *ApplicationCommandPermissionsList) (err error)
+	ApplicationCommands(appID, guildID string) (cmd []*ApplicationCommand, err error)
+	ApplicationCreate(ap *Application) (st *Application, err error)
+	ApplicationDelete(appID string) (err error)
+	ApplicationUpdate(appID string, ap *Application) (st *Application, err error)
+	Applications() (st []*Application, err error)
+	AutoModerationRule(guildID, ruleID string) (st *AutoModerationRule, err error)
+	AutoModerationRuleCreate(guildID string, rule *AutoModerationRule) (st *AutoModerationRule, err error)
+	AutoModerationRuleDelete(guildID, ruleID string) (err error)
+	AutoModerationRuleEdit(guildID, ruleID string, rule *AutoModerationRule) (st *AutoModerationRule, err error)
+	AutoModerationRules(guildID string) (st []*AutoModerationRule, err error)
+	Channel(channelID string) (st *Channel, err error)
+	ChannelDelete(channelID string) (st *Channel, err error)
+	ChannelEdit(channelID string, data *ChannelEdit) (st *Channel, err error)
+	ChannelEditComplex(channelID string, data *ChannelEdit) (st *Channel, err error)
+	ChannelFileSend(channelID, name string, r io.Reader) (*Message, error)
+	ChannelFileSendWithMessage(channelID, content string, name string, r io.Reader) (*Message, error)
+	ChannelInviteCreate(channelID string, i Invite) (st *Invite, err error)
+	ChannelInvites(channelID string) (st []*Invite, err error)
+	ChannelMessage(channelID, messageID string) (st *Message, err error)
+	ChannelMessageCrosspost(channelID, messageID string) (st *Message, err error)
+	ChannelMessageDelete(channelID, messageID string) (err error)
+	ChannelMessageEdit(channelID, messageID, content string) (*Message, error)
+	ChannelMessageEditComplex(m *MessageEdit) (st *Message, err error)
+	ChannelMessageEditEmbed(channelID, messageID string, embed *MessageEmbed) (*Message, error)
+	ChannelMessageEditEmbeds(channelID, messageID string, embeds []*MessageEmbed) (*Message, error)
+	ChannelMessagePin(channelID, messageID string) (err error)
+	ChannelMessageSend(channelID string, content string) (*Message, error)
+	ChannelMessageSendComplex(channelID string, data *MessageSend) (st *Message, err error)
+	ChannelMessageSendEmbed(channelID string, embed *MessageEmbed) (*Message, error)
+	ChannelMessageSendEmbedReply(channelID string, embed *MessageEmbed, reference *MessageReference) (*Message, error)
+	ChannelMessageSendEmbeds(channelID string, embeds []*MessageEmbed) (*Message, error)
+	ChannelMessageSendEmbedsReply(channelID string, embeds []*MessageEmbed, reference *MessageReference) (*Message, error)
+	ChannelMessageSendReply(channelID string, content string, reference *MessageReference) (*Message, error)
+	ChannelMessageSendTTS(channelID string, content string) (*Message, error)
+	ChannelMessageUnpin(channelID, messageID string) (err error)
+	ChannelMessages(channelID string, limit int, beforeID, afterID, aroundID string) (st []*Message, err error)
+	ChannelMessagesBulkDelete(channelID string, messages []string) (err error)
+	ChannelMessagesPinned(channelID string) (st []*Message, err error)
+	ChannelNewsFollow(channelID, targetID string) (st *ChannelFollow, err error)
+	ChannelPermissionDelete(channelID, targetID string) (err error)
+	ChannelPermissionSet(channelID, targetID string, targetType PermissionOverwriteType, allow, deny int64) (err error)
+	ChannelTyping(channelID string) (err error)
+	ChannelWebhooks(channelID string) (st []*Webhook, err error)
+	FollowupMessageCreate(interaction *Interaction, wait bool, data *WebhookParams) (*Message, error)
+	FollowupMessageDelete(interaction *Interaction, messageID string) error
+	FollowupMessageEdit(interaction *Interaction, messageID string, data *WebhookEdit) (*Message, error)
+	Gateway() (gateway string, err error)
+	GatewayBot() (st *GatewayBotResponse, err error)
+	Guild(guildID string) (st *Guild, err error)
+	GuildApplicationCommandsPermissions(appID, guildID string) (permissions []*GuildApplicationCommandPermissions, err error)
+	GuildAuditLog(guildID, userID, beforeID string, actionType, limit int) (st *GuildAuditLog, err error)
+	GuildBan(guildID, userID string) (st *GuildBan, err error)
+	GuildBanCreate(guildID, userID string, days int) (err error)
+	GuildBanCreateWithReason(guildID, userID, reason string, days int) (err error)
+	GuildBanDelete(guildID, userID string) (err error)
+	GuildBans(guildID string, limit int, beforeID, afterID string) (st []*GuildBan, err error)
+	GuildChannelCreate(guildID, name string, ctype ChannelType) (st *Channel, err error)
+	GuildChannelCreateComplex(guildID string, data GuildChannelCreateData) (st *Channel, err error)
+	GuildChannels(guildID string) (st []*Channel, err error)
+	GuildChannelsReorder(guildID string, channels []*Channel) (err error)
+	GuildCreate(name string) (st *Guild, err error)
+	GuildCreateWithTemplate(templateCode, name, icon string) (st *Guild, err error)
+	GuildDelete(guildID string) (st *Guild, err error)
+	GuildEdit(guildID string, g *GuildParams) (st *Guild, err error)
+	GuildEmbed(guildID string) (st *GuildEmbed, err error)
+	GuildEmbedEdit(guildID string, data *GuildEmbed) (err error)
+	GuildEmoji(guildID, emojiID string) (emoji *Emoji, err error)
+	GuildEmojiCreate(guildID string, data *EmojiParams) (emoji *Emoji, err error)
+	GuildEmojiDelete(guildID, emojiID string) (err error)
+	GuildEmojiEdit(guildID, emojiID string, data *EmojiParams) (emoji *Emoji, err error)
+	GuildEmojis(guildID string) (emoji []*Emoji, err error)
+	GuildIcon(guildID string) (img image.Image, err error)
+	GuildIntegrationCreate(guildID, integrationType, integrationID string) (err error)
+	GuildIntegrationDelete(guildID, integrationID string) (err error)
+	GuildIntegrationEdit(guildID, integrationID string, expireBehavior, expireGracePeriod int, enableEmoticons bool) (err error)
+	GuildIntegrations(guildID string) (st []*Integration, err error)
+	GuildInvites(guildID string) (st []*Invite, err error)
+	GuildLeave(guildID string) (err error)
+	GuildMember(guildID, userID string) (st *Member, err error)
+	GuildMemberAdd(guildID, userID string, data *GuildMemberAddParams) (err error)
+	GuildMemberDeafen(guildID string, userID string, deaf bool) (err error)
+	GuildMemberDelete(guildID, userID string) (err error)
+	GuildMemberDeleteWithReason(guildID, userID, reason string) (err error)
+	GuildMemberEdit(guildID, userID string, data *GuildMemberParams) (st *Member, err error)
+	GuildMemberEditComplex(guildID, userID string, data *GuildMemberParams) (st *Member, err error)
+	GuildMemberMove(guildID string, userID string, channelID *string) (err error)
+	GuildMemberMute(guildID string, userID string, mute bool) (err error)
+	GuildMemberNickname(guildID, userID, nickname string) (err error)
+	GuildMemberRoleAdd(guildID, userID, roleID string) (err error)
+	GuildMemberRoleRemove(guildID, userID, roleID string) (err error)
+	GuildMemberTimeout(guildID string, userID string, until *time.Time) (err error)
+	GuildMembers(guildID string, after string, limit int) (st []*Member, err error)
+	GuildMembersSearch(guildID, query string, limit int) (st []*Member, err error)
+	GuildPreview(guildID string) (st *GuildPreview, err error)
+	GuildPrune(guildID string, days uint32) (count uint32, err error)
+	GuildPruneCount(guildID string, days uint32) (count uint32, err error)
+	GuildRoleCreate(guildID string, data *RoleParams) (st *Role, err error)
+	GuildRoleDelete(guildID, roleID string) (err error)
+	GuildRoleEdit(guildID, roleID string, data *RoleParams) (st *Role, err error)
+	GuildRoleReorder(guildID string, roles []*Role) (st []*Role, err error)
+	GuildRoles(guildID string) (st []*Role, err error)
+	GuildScheduledEvent(guildID, eventID string, userCount bool) (st *GuildScheduledEvent, err error)
+	GuildScheduledEventCreate(guildID string, event *GuildScheduledEventParams) (st *GuildScheduledEvent, err error)
+	GuildScheduledEventDelete(guildID, eventID string) (err error)
+	GuildScheduledEventEdit(guildID, eventID string, event *GuildScheduledEventParams) (st *GuildScheduledEvent, err error)
+	GuildScheduledEventUsers(guildID, eventID string, limit int, withMember bool, beforeID, afterID string) (st []*GuildScheduledEventUser, err error)
+	GuildScheduledEvents(guildID string, userCount bool) (st []*GuildScheduledEvent, err error)
+	GuildSplash(guildID string) (img image.Image, err error)
+	GuildTemplate(templateCode string) (st *GuildTemplate, err error)
+	GuildTemplateCreate(guildID string, data *GuildTemplateParams) (st *GuildTemplate)
+	GuildTemplateDelete(guildID, templateCode string) (err error)
+	GuildTemplateEdit(guildID, templateCode string, data *GuildTemplateParams) (st *GuildTemplate, err error)
+	GuildTemplateSync(guildID, templateCode string) (err error)
+	GuildTemplates(guildID string) (st []*GuildTemplate, err error)
+	GuildThreadsActive(guildID string) (threads *ThreadsList, err error)
+	GuildWebhooks(guildID string) (st []*Webhook, err error)
+	GuildWithCounts(guildID string) (st *Guild, err error)
+	InteractionRespond(interaction *Interaction, resp *InteractionResponse) error
+	InteractionResponse(interaction *Interaction) (*Message, error)
+	InteractionResponseDelete(interaction *Interaction) error
+	InteractionResponseEdit(interaction *Interaction, newresp *WebhookEdit) (*Message, error)
+	Invite(inviteID string) (st *Invite, err error)
+	InviteAccept(inviteID string) (st *Invite, err error)
+	InviteComplex(inviteID, guildScheduledEventID string, withCounts, withExpiration bool) (st *Invite, err error)
+	InviteDelete(inviteID string) (st *Invite, err error)
+	InviteWithCounts(inviteID string) (st *Invite, err error)
+	MessageReactionAdd(channelID, messageID, emojiID string) error
+	MessageReactionRemove(channelID, messageID, emojiID, userID string) error
+	MessageReactions(channelID, messageID, emojiID string, limit int, beforeID, afterID string) (st []*User, err error)
+	MessageReactionsRemoveAll(channelID, messageID string) error
+	MessageReactionsRemoveEmoji(channelID, messageID, emojiID string) error
+	MessageThreadStart(channelID, messageID string, name string, archiveDuration int) (ch *Channel, err error)
+	MessageThreadStartComplex(channelID, messageID string, data *ThreadStart) (ch *Channel, err error)
+	Request(method, urlStr string, data interface{}) (response []byte, err error)
+	RequestWithBucketID(method, urlStr string, data interface{}, bucketID string) (response []byte, err error)
+	RequestWithLockedBucket(method, urlStr, contentType string, b []byte, bucket *Bucket, sequence int) (response []byte, err error)
+	StageInstance(channelID string) (si *StageInstance, err error)
+	StageInstanceCreate(data *StageInstanceParams) (si *StageInstance, err error)
+	StageInstanceDelete(channelID string) (err error)
+	StageInstanceEdit(channelID string, data *StageInstanceParams) (si *StageInstance, err error)
+	ThreadJoin(id string) error
+	ThreadLeave(id string) error
+	ThreadMember(threadID, memberID string) (member *ThreadMember, err error)
+	ThreadMemberAdd(threadID, memberID string) error
+	ThreadMemberRemove(threadID, memberID string) error
+	ThreadMembers(threadID string) (members []*ThreadMember, err error)
+	ThreadStart(channelID, name string, typ ChannelType, archiveDuration int) (ch *Channel, err error)
+	ThreadStartComplex(channelID string, data *ThreadStart) (ch *Channel, err error)
+	ThreadsActive(channelID string) (threads *ThreadsList, err error)
+	ThreadsArchived(channelID string, before *time.Time, limit int) (threads *ThreadsList, err error)
+	ThreadsPrivateArchived(channelID string, before *time.Time, limit int) (threads *ThreadsList, err error)
+	ThreadsPrivateJoinedArchived(channelID string, before *time.Time, limit int) (threads *ThreadsList, err error)
+	User(userID string) (st *User, err error)
+	UserAvatar(userID string) (img image.Image, err error)
+	UserAvatarDecode(u *User) (img image.Image, err error)
+	UserChannelCreate(recipientID string) (st *Channel, err error)
+	UserChannelPermissions(userID, channelID string) (apermissions int64, err error)
+	UserConnections() (conn []*UserConnection, err error)
+	UserGuildMember(guildID string) (st *Member, err error)
+	UserGuilds(limit int, beforeID, afterID string) (st []*UserGuild, err error)
+	UserUpdate(username, avatar string) (st *User, err error)
+	VoiceRegions() (st []*VoiceRegion, err error)
+	Webhook(webhookID string) (st *Webhook, err error)
+	WebhookCreate(channelID, name, avatar string) (st *Webhook, err error)
+	WebhookDelete(webhookID string) (err error)
+	WebhookDeleteWithToken(webhookID, token string) (st *Webhook, err error)
+	WebhookEdit(webhookID, name, avatar, channelID string) (st *Role, err error)
+	WebhookEditWithToken(webhookID, token, name, avatar string) (st *Role, err error)
+	WebhookExecute(webhookID, token string, wait bool, data *WebhookParams) (st *Message, err error)
+	WebhookMessage(webhookID, token, messageID string) (message *Message, err error)
+	WebhookMessageDelete(webhookID, token, messageID string) (err error)
+	WebhookMessageEdit(webhookID, token, messageID string, data *WebhookEdit) (st *Message, err error)
+	WebhookThreadExecute(webhookID, token string, wait bool, threadID string, data *WebhookParams) (st *Message, err error)
+	WebhookWithToken(webhookID, token string) (st *Webhook, err error)
+}
